+++ src/screens/Documents/Documents.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TextRegular } from '../../assets/fonts/Fonts';
+import { TextRegular } from '../../../assets/fonts/Fonts';
 import { PageContainer } from '../styles';
 
 const UploadScreen = () => {
