+++ src/screens/Welcome/styles.ts
