+++ src/screens/auth/LoginScreen.tsx
@@ -13,10 +13,10 @@ import {
   TextRegularWhite,
   TextRegularRed,
   TextSubtitle,
-} from '../../assets/fonts/Fonts';
-import { ButtonDark, TextInput } from '../../assets/Components';
+} from '../../../assets/fonts/Fonts';
+import { ButtonDark, TextInput } from '../../../assets/Components';
 import { PageContainer } from '../styles';
-import { login } from '../../firebase/auth';
+import { login } from '../../database/auth';
 
 const LoginScreen = ({ navigation }: any) => {
   const [email, setEmail] = useState('');
