+++ src/screens/auth/styles.ts
